Django/MyBlog(Developed Own - Few Helps Only)/requirements.txt
asgiref-3.4.1 
backports.zoneinfo-0.2.1 
django-4.0 
sqlparse-0.4.2 
tzdata-2021.5